move inactiveExtensionUrlHandler to common
