+++ src/vs/workbench/services/extensions/common/extensionDescriptionRegistry.ts
