+++ src/vs/workbench/services/extensions/common/extensionHostProtocol.ts
