+++ src/vs/workbench/services/extensions/common/inactiveExtensionUrlHandler.ts
