+++ src/vs/workbench/services/extensions/common/lazyPromise.ts
