+++ src/vs/workbench/services/extensions/common/proxyIdentifier.ts
